+++ src/plugins/three/fade/FadeBatchedMesh.js
@@ -1,6 +1,5 @@
 import { PassThroughBatchedMesh } from './PassThroughBatchedMesh.js';
-import { InstanceDataTexture } from './InstanceDataTexture.js';
-import { RGFormat, UnsignedByteType } from 'three';
+import { RGFormat, UnsignedByteType, DataTexture } from 'three';
 import { wrapFadeMaterial } from './wrapFadeMaterial.js';
 
 export class FadeBatchedMesh extends PassThroughBatchedMesh {
@@ -9,20 +8,20 @@ export class FadeBatchedMesh extends PassThroughBatchedMesh {
 
 		super( ...args );
 
+		// construct a version of the material that supports fading
 		const material = this.material;
 		const params = wrapFadeMaterial( material, material.onBeforeCompile );
 		material.defines.FEATURE_FADE = 1;
 		material.defines.USE_BATCHING_FRAG = 1;
 		material.needsUpdate = true;
 
+		// fade parameters
 		this.fadeTexture = null;
 		this._fadeParams = params;
 
-		this._initFadeTexture();
-
-
 	}
 
+	// Set the fade state
 	setFadeAt( index, fadeIn, fadeOut ) {
 
 		this._initFadeTexture();
@@ -30,19 +29,25 @@ export class FadeBatchedMesh extends PassThroughBatchedMesh {
 
 	}
 
+	// initialize the texture and resize it if needed
 	_initFadeTexture() {
 
+		// calculate the new size
 		let size = Math.sqrt( this._maxInstanceCount );
 		size = Math.ceil( size );
 
+		const length = size * size * 2;
 		const oldFadeTexture = this.fadeTexture;
-		if ( ! this.fadeTexture || this.fadeTexture.image.data.length !== size * size * 2 ) {
+		if ( ! oldFadeTexture || oldFadeTexture.image.data.length !== length ) {
 
-			// 4 floats per RGBA pixel initialized to white
-			const fadeArray = new Uint8Array( size * size * 2 );
+			// 2 bytes per RG pixel
+			const fadeArray = new Uint8Array( length );
 			const fadeTexture = new InstanceDataTexture( fadeArray, size, size, RGFormat, UnsignedByteType );
 
+			// copy the data from the old fade texture if it exists
 			if ( oldFadeTexture ) {
+
+				oldFadeTexture.dispose();
 
 				const src = oldFadeTexture.image.data;
 				const dst = this.fadeTexture.image.data;
@@ -51,6 +56,7 @@ export class FadeBatchedMesh extends PassThroughBatchedMesh {
 
 			}
 
+			// assign the new fade texture to the uniform, member variable
 			this.fadeTexture = fadeTexture;
 			this._fadeParams.fadeTexture.value = fadeTexture;
 			fadeTexture.needsUpdate = true;
@@ -59,9 +65,42 @@ export class FadeBatchedMesh extends PassThroughBatchedMesh {
 
 	}
 
+	// dispose the fade texture. Super cannot be used here due to proxy
 	dispose() {
 
 		this.fadeTexture.dispose();
+
+	}
+
+}
+
+// Version of data texture that can assign pixel values
+class InstanceDataTexture extends DataTexture {
+
+	setValueAt( instance, ...values ) {
+
+		const { data, width, height } = this.image;
+		const itemSize = Math.floor( data.length / ( width * height ) );
+		let needsUpdate = false;
+		for ( let i = 0; i < itemSize; i ++ ) {
+
+			const index = instance * itemSize + i;
+			const prevValue = data[ index ];
+			const newValue = values[ i ] || 0;
+			if ( prevValue !== newValue ) {
+
+				data[ index ] = newValue;
+				needsUpdate = true;
+
+			}
+
+		}
+
+		if ( needsUpdate ) {
+
+			this.needsUpdate = true;
+
+		}
 
 	}
